integrate cmd package and change test to without output
